Added lazy_f variant.
The original attempt at just making 'f'
lazy doesn't quite work: it changes
the return type, which is then blows
up in all kinds of ways in the unit-tests.

This version introduces another variant
that evaluates on the first conversion
to 'str', then caches the result.

Some unit-tests are added to test
scope-preservation and caching
behavior.